+++ CHANGELOG.rst
@@ -5,7 +5,8 @@ Version History
 v6.4.0
 ------
 
-* Add M1M3Compact and M2Compact components `<https://github.com/lsst-ts/LOVE-frontend/pull/673>`_
+* Add M1M3HardpointsDataTable component. `<https://github.com/lsst-ts/LOVE-frontend/pull/674>`_
+* Add M1M3Compact and M2Compact components. `<https://github.com/lsst-ts/LOVE-frontend/pull/673>`_
 
 v6.3.2
 ------